making components and deleting coments
