+++ src/lib/apiKey.js
@@ -1,5 +1,3 @@
-//import ApiKey from "../views/apiKey";
-
 export const getApiKey = () => {
   const valor = localStorage.getItem("ApiKey");
   return valor;
